--- a/src/CircularArrays.jl
+++ b/src/CircularArrays.jl
@@ -10,7 +10,7 @@ export CircularArray, CircularVector
 
 `N`-dimensional array backed by an `AbstractArray{T, N}` of type `A` with fixed size and circular indexing.
 
-    array[index] == array[mod1(index, size)]
+    array[index...] == array[mod1.(index, size)...]
 """
 struct CircularArray{T, N, A} <: AbstractArray{T, N}
     data::A
@@ -23,7 +23,7 @@ end
 One-dimensional array backed by an `AbstractArray{T, 1}` of type `A` with fixed size and circular indexing.
 Alias for [`CircularArray{T,1,A}`](@ref).
 
-    array[index] == array[mod1(index, size)]
+    array[index] == array[mod1(index, length)]
 """
 const CircularVector{T} = CircularArray{T, 1}
 
@@ -34,12 +34,12 @@ CircularArray(def::T, size) where T = CircularArray(fill(def, size))
 @inline Base.getindex(arr::CircularArray, i::Integer) =
     @inbounds getindex(arr.data, mod(i, eachindex(LinearIndices(arr.data))))
 @inline Base.getindex(arr::CircularArray{T,N}, I::Vararg{<:Integer,N}) where {T,N} =
-    @inbounds getindex(arr.data, mod.(I, axes(arr))...)
+    @inbounds getindex(arr.data, map(mod, I, axes(arr.data))...)
 
 @inline Base.setindex!(arr::CircularArray, v, i::Integer) =
     @inbounds setindex!(arr.data, v, mod(i, eachindex(LinearIndices(arr.data))))
 @inline Base.setindex!(arr::CircularArray{T,N}, v, I::Vararg{<:Integer,N}) where {T,N} =
-    @inbounds setindex!(arr.data, v, mod.(I, axes(arr))...)
+    @inbounds setindex!(arr.data, v, map(mod, I, axes(arr.data))...)
 
 @inline Base.size(arr::CircularArray) = size(arr.data)
 @inline Base.axes(arr::CircularArray) = axes(arr.data)
@@ -62,5 +62,12 @@ CircularVector(def::T, size::Int) where T = CircularVector{T}(fill(def, size))
 
 Base.IndexStyle(::Type{CircularArray{T,N,A}}) where {T,N,A} = IndexStyle(A)
 Base.IndexStyle(::Type{<:CircularVector}) = IndexLinear()
+
+function Base.showarg(io::IO, arr::CircularArray, toplevel)
+    print(io, ndims(arr) == 1 ? "CircularVector(" : "CircularArray(")
+    Base.showarg(io, parent(arr), false)
+    print(io, ')')
+    # toplevel && print(io, " with eltype ", eltype(arr))
+end
 
 end