--- a/src/CircularArrays.jl
+++ b/src/CircularArrays.jl
@@ -35,11 +35,15 @@ CircularArray(def::T, size) where T = CircularArray(fill(def, size))
     @inbounds getindex(arr.data, mod1(i, length(arr.data)))
 @inline Base.getindex(arr::CircularArray{T,N}, I::Vararg{<:Int,N}) where {T,N} =
     @inbounds getindex(arr.data, map(mod, I, axes(arr.data))...)
+@inline Base._getindex(::IndexLinear, arr::CircularArray{T,N}, I::Vararg{<:Int,N}) where {T,N} =
+    @inbounds Base._getindex(IndexLinear(), arr.data, map(mod, I, axes(arr.data))...)
 
 @inline Base.setindex!(arr::CircularArray, v, i::Int) =
     @inbounds setindex!(arr.data, v, mod1(i, length(arr.data)))
 @inline Base.setindex!(arr::CircularArray{T,N}, v, I::Vararg{<:Int,N}) where {T,N} =
     @inbounds setindex!(arr.data, v, map(mod, I, axes(arr.data))...)
+@inline Base._setindex!(::IndexLinear, arr::CircularArray{T,N}, v, I::Vararg{<:Int,N}) where {T,N} =
+    @inbounds Base._setindex!(IndexLinear(), arr.data, v, map(mod, I, axes(arr.data))...)
 
 @inline Base.size(arr::CircularArray) = size(arr.data)
 @inline Base.axes(arr::CircularArray) = axes(arr.data)