--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -79,6 +79,9 @@ end
     a2 = CircularArray(4, (2, 3))
     @test isa(a2, CircularArray{Int, 2})
 
+    @test a1[3] == a1[3,1] # linear indexing
+    @test a1[4] == a1[1,2]
+
     @test a1[2, 3, 1] == 17 # trailing index
     @test a1[2, 3, 99] == 17
     @test a1[2, 3, :] == [17]