Before:
```
julia> ca = CircularArray(rand(Int8, 10,10));

julia> @btime [$ca[i,j] for i in 1:21, j in 1:21];
  52.352 μs (886 allocations: 48.91 KiB)
```
After:
```
julia> @btime [$ca[i,j] for i in 1:21, j in 1:21]
  4.009 μs (1 allocation: 544 bytes)
21×21 Matrix{Int8}:
 -69  -22   -83  -28    -7   -20   106  -119  …  -28    -7   -20   106  -119    59  -111  -69
  25  -72  -123  118    36   -54   -81   118     118    36   -54   -81   118   -90   -77   25
...
```