@mcabbott 2020-11-16
One behaviour changed by this PR is:
```
julia> a1 = CircularArray(b_arr)
3×4 CircularArray{Int64, 2, Matrix{Int64}}:
  2   4   6   8
 10  12  14  16
 18  20  22  24

julia> a1[3]
18

julia> a1[4] # was == 2
4
```
So linear indexing will cover the whole matrix, not just the first column. It still has ` IndexStyle(a1) == IndexCartesian()` but could inherit that from the parent, too.