@Vexatos 2020-11-17
Thanks for the pull request! Here are some performance results from me.
```
Before:

julia> ca = CircularArray(rand(Int8, 10,10));

julia> @btime [$ca[i,j] for i in 1:21, j in 1:21];
  43.500 μs (883 allocations: 48.77 KiB)

julia> @btime [$ca[i] for i in 1:(21*21)]
  2.125 μs (1 allocation: 544 bytes)

julia> ca = CircularArray(rand(Int8, 10));

julia> @btime [$ca[i] for i in 1:(21*21)];
  2.122 μs (1 allocation: 544 bytes)


After:

julia> ca = CircularArray(rand(Int8, 10,10));

julia> @btime [$ca[i,j] for i in 1:21, j in 1:21];
  4.855 μs (1 allocation: 544 bytes)

julia> @btime [$ca[i] for i in 1:(21*21)]
  5.965 μs (1 allocation: 544 bytes)

julia> ca = CircularArray(rand(Int8, 10));

julia> @btime [$ca[i] for i in 1:(21*21)];
  2.124 μs (1 allocation: 544 bytes)

```

Linear indexing of linear arrays didn't change, 2D-indexing of matrices got faster just as you reported, but linear indexing of matrices has gotten slower. All these results can be reproduced for larger arrays (200×200 elements accessed) and more dimensions. Linear indexing of a higher-dimensional array is consistently about 3 times slower with the PR than without for 2D arrays, 4 times slower with 3D arrays and 6 times slower with 4D arrays.

```
Before:

julia> ca = CircularArray(rand(Int8, 4, 4, 4));

julia> @btime [$ca[i] for i in 1:(21*21)];
  2.010 μs (1 allocation: 544 bytes)

julia> ca = CircularArray(rand(Int8, 3, 3, 3, 3));

julia> @btime [$ca[i] for i in 1:(21*21)];
  2.123 μs (1 allocation: 544 bytes)

After:

julia> ca = CircularArray(rand(Int8, 4, 4, 4));

julia> @btime [$ca[i] for i in 1:(21*21)];
  10.064 μs (1 allocation: 544 bytes)

julia> ca = CircularArray(rand(Int8, 3, 3, 3, 3));

julia> @btime [$ca[i] for i in 1:(21*21)];
  13.392 μs (1 allocation: 544 bytes)
```

Linear indexing of the array has so far been undocumented behaviour, so changing the behaviour doesn't seem to be much of a problem to me, but the slowdown getting worse the more dimensions the array has is worrying. That should probably be addressed.

Adding `Base.parent` is a good idea, thanks for that.