@Vexatos 2020-11-17
If I had to guess, this slowdown may be caused by the linear indexing special case being removed, making linear and multi-dimensional indexing use the same method. In the old version, the special case for linear indexing did not use a splatting operator `...`, which from my experience is very slow and might be the source of this slowdown during linear indexing.

Furthermore, I discovered that the arrays no longer error on indexing with bad dimensions:

Before:
```
julia> ca = CircularArray(rand(Int8, 3, 3, 3, 3));

julia> ca[1, 2]
ERROR: BoundsError: attempt to access 3×3×3×3 Array{Int8,4} at index [1, 2]
Stacktrace:
 [1] getindex(::Array{Int8,4}, ::Int64, ::Int64) at ./array.jl:810
 [2] getindex(::CircularArray{Int8,4,Array{Int8,4}}, ::Int64, ::Int64) at /home/durian/Dokumente/Development/Other/CircularArrays.jl/src/CircularArrays.jl:38
 [3] top-level scope at REPL[6]:1
```
After:
```
julia> ca = CircularArray(rand(Int8, 3, 3, 3, 3));

julia> ca[1, 2]
75						# equal to ca[1, 2, 1, 1]
```