@mcabbott 2020-11-17
I'm not sure we should call the change in time for linear indexing of a matrix a slowdown, as it's doing a different operation. Instead of accessing the first column repeatedly, it reads out all the data, and either your calculation wants one or it wants the other. But I am not surprised that reading all the data takes longer, I'd point to thinks fitting in registers or cache or something like that. 

And, to go from Cartesian to Linear indices is just multiplication, but the reverse involves `div` which is slower. And there's more `div`ing in higher dims. For example, compare an Array to a trivial PermutedDimsArray just to fool linear indexing:
```
julia> quad = randn(10,10,10,10);

julia> IndexStyle(quad)
IndexLinear()

julia> @btime [$quad[i] for i in 1:10^4];
  7.216 μs (2 allocations: 78.20 KiB)

julia> pquad = PermutedDimsArray(quad, (1,2,3,4));

julia> IndexStyle(pquad)
IndexCartesian()

julia> @btime [$pquad[i] for i in 1:10^4];
  22.557 μs (2 allocations: 78.20 KiB)
```
The old behaviour didn't pay this cost, because it fixed all indices but the leftmost one to 1. 

What is still weird about linear indexing is that the algorithm from Base it's using to convert to Cartesian seems not to behave well with negative numbers. 
```
julia> a1
3×4 CircularArray{Int64, 2, Matrix{Int64}}:
  2   4   6   8
 10  12  14  16
 18  20  22  24

julia> a1[-2:9]' # after this PR
1×12 LinearAlgebra.Adjoint{Int64, CircularVector{Int64, Vector{Int64}}}:
 8  10  18  2  10  18  4  12  20  6  14  22

julia> a1[-2], a1[9]
(8, 22)

julia> CircularArray(1:3)[-2:6]'  # with a vector, no conversion to cartesian, no problem:
1×9 LinearAlgebra.Adjoint{Int64, CircularVector{Int64, Vector{Int64}}}:
 1  2  3  1  2  3  1  2  3
```
Before was even weirder, I think:
```
julia> a1[-2:9]' 
1×12 LinearAlgebra.Adjoint{Int64, CircularVector{Int64, Vector{Int64}}}:
 8  10  18  2  10  18  4  12  20  6  14  22

julia> a1[-2], a1[9]  # doesn't match
(2, 18)
```