@mcabbott 2020-11-17
> Furthermore, I discovered that the arrays no longer error on indexing with bad dimensions:

Oh I didn't see that variant. I saw this one, and added a test:
```
julia> a1[2,3,99]
14

julia> size(a1, 3)
1
```
I'm not sure how weird this is, given that ordinary arrays do handle things like this:
```
julia> cquad[3,4] == quad[3,4,1,1] == quad[3,4,1,1,1,1,1,1]
true
```
But if you like I'll restore the error.