@mcabbott 2020-11-17
But when the underlying array supports fast linear indexing, then there's no need to go via Cartesian. So the latest commit speeds this up:
```
julia> @btime [$ca[i,j] for i in 1:21, j in 1:21];
  4.015 μs (1 allocation: 544 bytes)

julia> @btime [$ca[i] for i in 1:(21*21)];
  1.503 μs (1 allocation: 544 bytes)
```
and avoids some weirdness:
```
julia> a1
3×4 CircularArray{Int64, 2, Matrix{Int64}}:
  2   4   6   8
 10  12  14  16
 18  20  22  24

julia> a1[-2:9]'
1×12 LinearAlgebra.Adjoint{Int64, CircularVector{Int64, Vector{Int64}}}:
 8  16  24  2  10  18  4  12  20  6  14  22

julia> a1[-2], a1[9]
(8, 22)
```
and adds a bounds check to restore the error on too few indices. 

But breaks one offset array test, not sure why yet.